Update db (removed wallet_id from passenger)
